Creates input and response components to tidy files
